backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
fastapi
uvicorn
python-dotenv
requests